Dealing with large, unfiltered files
Added a new pre-filtering script to go through files too big to process
in memory.  This has the advantage of preventing memory errors, but the
disadvantage of often taking as long as all the other steps combines.
I will probably move this specific part into a compiles language
(either C++ or Go, since I’m trying to learn it now) and see if I can
squeeze a little more performance out of it.